requirements.txt
pymongo==4.2.0

python-dotenv~=0.21.0